Created new files for Colors & Constants with minor code clean up and warning fixes
