+++ SnapList/SNColors.swift
@@ -1,0 +1,12 @@
+//
+//  SNColors.swift
+//  SnapList
+//
+//  Created by Aamir  on 28/08/17.
+//  Copyright © 2017 AamirAnwar. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+let SNLightGray = UIColor.init(hex: 0xF1F1F1)
