+++ SnapList/SNConstants.swift
@@ -1,0 +1,21 @@
+//
+//  SNConstants.swift
+//  SnapList
+//
+//  Created by Aamir  on 28/08/17.
+//  Copyright © 2017 AamirAnwar. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+let kpadding13:CGFloat = 13.0
+let basePath = "https://snaplist-server.herokuapp.com/api"
+let NotificationUserLoggedSuccessfully = "userLoggedInSuccessfully"
+let KeyUserID = "USER_ID"
+let KeyListID = "LIST_ID"
+let PADDING_8:CGFloat = 8
+let STATUS_BAR_HEIGHT:CGFloat = 22
+let kButtonHeight:CGFloat = 50
+let KeyShareCode = "SHARE_CODE"
+let NotificationItemAdded = "NotificationItemAdded"
