+++ SnapList/View Controllers/SNInviteViewController.swift
@@ -14,12 +14,5 @@ class SNInviteViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         shareCodeLabel.text = UserDefaults.standard.value(forKey: KeyShareCode) as? String
-        // Do any additional setup after loading the view.
     }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
 }
